backend/requirements.txt
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
better-exceptions==0.2.2
boto3==1.13.4
botocore==1.16.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-model-utils==4.0.0
django-safedelete==0.5.4
djangorestframework==3.11.0
docutils==0.15.2
dry-rest-permissions==0.1.10
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
olefile==0.46
pep8==1.7.1
Pillow==7.1.2
pycodestyle==2.5.0
pycparser==2.20
pylint==2.5.2
pyOpenSSL==19.1.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
sendgrid==6.3.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
